requirements.txt
fastapi~=0.78.0
uvicorn~=0.18.2
requests~=2.28.0
jinja2
pydantic~=1.9.1
uuid~=1.30
passlib~=1.7.4
starlette~=0.19.1
bcrypt
python-multipart
gunicorn
aiofiles
redis~=4.3.4
psycopg2~=2.9.3
SQLAlchemy~=1.4.39
python-dotenv~=0.20.0